refactor: move oidc validator mocks into pkg/auth/mocks
